+++ src/main/kotlin/lotto/util/Reader.kt
@@ -4,9 +4,9 @@ import camp.nextstep.edu.missionutils.Console
 
 object Reader {
     private fun readLine() = Console.readLine()
-    private const val RETURN_CODE = 0L
+    private const val RETURN_CODE = 0
     private val RETURN_LIST = listOf(0)
-    fun readLottoMoney(): Long {
+    fun readLottoMoney(): Int {
         var money = RETURN_CODE
         while (money == RETURN_CODE) {
             money = InputChecker.checkInputMoney(readLine(), RETURN_CODE)
@@ -27,9 +27,9 @@ object Reader {
     }
 
     fun readLottoBonus(lottoNumbers: List<Int>): Int {
-        var bonus = RETURN_CODE.toInt()
-        while (bonus == RETURN_CODE.toInt()) {
-            bonus = InputChecker.checkInputBonus(lottoNumbers, readLine(), RETURN_CODE.toInt())
+        var bonus = RETURN_CODE
+        while (bonus == RETURN_CODE) {
+            bonus = InputChecker.checkInputBonus(lottoNumbers, readLine(), RETURN_CODE)
             println()
         }
